pyproject.toml
[tool.poetry]
name = "django_custom_user_model"
version = "0.1.0"
description = ""
authors = ["bashar adel taamneh <bashartaamneh55@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1.4"
django-allauth = "0.43.0"
django-crispy-forms = "1.10.0"
django-debug-toolbar = "3.2"
whitenoise = "5.2.0"
django-bootstrap-v5 = "^1.0.7"
Pillow = "^8.4.0"
sorl-thumbnail = "^12.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"